+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- markdownlint-disable -->
 ## [Unreleased]
 
+### Changed
+- Update `netlify-js-app`
+- Update rollup config files
+
 ## [v1.2.0] - 2023-05-12
 
 ### Added